Subscriptions (#5)
* Show user

* Added fields for companis

* Create stripe plans

* Create stripe plan

* Subscriptin done

+++ app/Imports/Sales/Sheets/InvoiceItemTaxes.php
@@ -26,18 +26,18 @@ class InvoiceItemTaxes extends Import
 
         $row['document_id'] = (int) Document::invoice()->number($row['invoice_number'])->pluck('id')->first();
 
-        if (empty($row['document_item_id']) && !empty($row['item_name'])) {
+        if (empty($row['document_item_id']) && ! empty($row['item_name'])) {
             $item_id = Item::name($row['item_name'])->pluck('id')->first();
             $row['document_item_id'] = DocumentItem::invoice()->where('item_id', $item_id)->pluck('id')->first();
         }
 
         $row['tax_id'] = $this->getTaxId($row);
 
-        if (empty($row['name']) && !empty($row['item_name'])) {
+        if (empty($row['name']) && ! empty($row['item_name'])) {
             $row['name'] = $row['item_name'];
         }
 
-        $row['amount'] = (double) $row['amount'];
+        $row['amount'] = (float) $row['amount'];
 
         $row['type'] = Document::INVOICE_TYPE;
